+++ src/auth/client.ts
@@ -1,9 +1,9 @@
 import { NodeOAuthClient } from '@atproto/oauth-client-node'
-import type { Database } from '#/db'
 import { env } from '#/lib/env'
 import { SessionStore, StateStore } from './storage'
+import { MongoClient } from 'mongodb'
 
-export const createClient = async (db: Database) => {
+export const createClient = async ( dbm: MongoClient) => {
   const publicUrl = env.PUBLIC_URL
   const url = publicUrl || `http://127.0.0.1:${env.PORT}`
   const enc = encodeURIComponent
@@ -22,7 +22,7 @@ export const createClient = async (db: Database) => {
       token_endpoint_auth_method: 'none',
       dpop_bound_access_tokens: true,
     },
-    stateStore: new StateStore(db),
-    sessionStore: new SessionStore(db),
+    stateStore: new StateStore(dbm),
+    sessionStore: new SessionStore(dbm),
   })
 }
