+++ src/auth/storage.ts
@@ -4,44 +4,61 @@ import type {
   NodeSavedState,
   NodeSavedStateStore,
 } from '@atproto/oauth-client-node'
-import type { Database } from '#/db'
+  import { MongoClient } from 'mongodb'
 
 export class StateStore implements NodeSavedStateStore {
-  constructor(private db: Database) {}
+  private db
+  private collection
+
+  constructor(private dbm: MongoClient) {
+    this.db = this.dbm.db('statusphere')
+    this.collection = this.db.collection('auth_state')
+  }
   async get(key: string): Promise<NodeSavedState | undefined> {
-    const result = await this.db.selectFrom('auth_state').selectAll().where('key', '=', key).executeTakeFirst()
+    const result = await this.collection.findOne({ key })
     if (!result) return
     return JSON.parse(result.state) as NodeSavedState
   }
+  
   async set(key: string, val: NodeSavedState) {
     const state = JSON.stringify(val)
-    await this.db
-      .insertInto('auth_state')
-      .values({ key, state })
-      .onConflict((oc) => oc.doUpdateSet({ state }))
-      .execute()
+    const result = await this.collection.insertOne({key: key, state: state})
   }
+  
   async del(key: string) {
-    await this.db.deleteFrom('auth_state').where('key', '=', key).execute()
+    const result = await this.collection.deleteOne({ key: key})
   }
 }
 
 export class SessionStore implements NodeSavedSessionStore {
-  constructor(private db: Database) {}
+  private db
+  private collection
+
+  constructor(private dbm: MongoClient) {
+    this.db = this.dbm.db('statusphere')
+    this.collection = this.db.collection('auth_session')
+  }
+  
   async get(key: string): Promise<NodeSavedSession | undefined> {
-    const result = await this.db.selectFrom('auth_session').selectAll().where('key', '=', key).executeTakeFirst()
+    const result = await this.collection.findOne({ key })
+    //console.log('session-get-result:', result)
     if (!result) return
     return JSON.parse(result.session) as NodeSavedSession
   }
+
   async set(key: string, val: NodeSavedSession) {
     const session = JSON.stringify(val)
-    await this.db
-      .insertInto('auth_session')
-      .values({ key, session })
-      .onConflict((oc) => oc.doUpdateSet({ session }))
-      .execute()
+    const result = await this.collection.updateOne(
+      { key: key },
+      { $set: { session: session } },
+      { upsert: true }
+    )
+    console.log('session-set-result:', result)
   }
+
   async del(key: string) {
-    await this.db.deleteFrom('auth_session').where('key', '=', key).execute()
+    console.log('session-del-key:', key)
+    const result = await this.collection.deleteOne({ key: key})
+    console.log('session-del-result:', result)
   }
 }
